webhooks appear to be working now

+++ Common/Extensions/StringExtensions.cs
@@ -1,5 +1,5 @@
 ﻿namespace Common.Extensions;
 public static class StringExtensions
 {
-    public static int? ToTeamNumber(this string? s) => int.TryParse(s?.TrimStart(['f', 'r', 'c']), out var i) ? i : null;
+    public static ushort? ToTeamNumber(this string? s) => ushort.TryParse(s?.TrimStart(['f', 'r', 'c']), out var i) ? i : null;
 }
